Complete refactoring of the repository structure
