+++ echemsuite/cellcycling/__init__.py
@@ -1,1 +1,2 @@
-from echemsuite.cellcycling.read_input import read_cycles, build_cycles, time_adjust
+from echemsuite.cellcycling.cycles import time_adjust
+from echemsuite.cellcycling.read_input import read_cycles, build_cycles
